attempt to sparsify B' output wherever possible

+++ core/src/test/java/org/apache/mahout/math/hadoop/stochasticsvd/LocalSSVDSolverSparseSequentialTest.java
@@ -88,7 +88,7 @@ public class LocalSSVDSolverSparseSequentialTest extends MahoutTestCase {
     int ablockRows = 867;
     int p = 60;
     int k = 40;
-    int percent = 5;
+    double percent = 5;
 
     VectorWritable vw = new VectorWritable();
     IntWritable roww = new IntWritable();
